Migrate seo utils
